Remove interfaces file
