--- packages/js/app/src/Interfaces/InfoProps.ts
@@ -1,9 +1,0 @@
-import { IconProp } from '@fortawesome/fontawesome-svg-core';
-
-interface InfoProps {
-    title: string;
-    detail: string;
-    icon: IconProp;
-}
-
-export default InfoProps;
