--- packages/js/app/src/Interfaces/ProfileImageProps.ts
@@ -1,6 +1,0 @@
-interface ProfileImageProps {
-    source: string;
-    size: number;
-}
-
-export default ProfileImageProps;
