--- packages/js/app/src/Interfaces/SlideCardProps.ts
@@ -1,9 +1,0 @@
-import { IconProp } from '@fortawesome/fontawesome-svg-core';
-
-interface SlideCardProps {
-    title: string;
-    detail: string;
-    icon: IconProp;
-}
-
-export default SlideCardProps;
